+++ packages/medusa/src/api/routes/admin/index.js
@@ -19,6 +19,7 @@ import appRoutes from "./apps"
 import swapRoutes from "./swaps"
 import returnRoutes from "./returns"
 import variantRoutes from "./variants"
+import draftOrderRoutes from "./draft-orders"
 import collectionRoutes from "./collections"
 import notificationRoutes from "./notifications"
 
@@ -63,6 +64,7 @@ export default (app, container, config) => {
   swapRoutes(route)
   returnRoutes(route)
   variantRoutes(route)
+  draftOrderRoutes(route)
   collectionRoutes(route)
   notificationRoutes(route)
   returnReasonRoutes(route)