+++ packages/medusa/src/api/routes/admin/orders/__tests__/get-order.js
@@ -41,6 +41,7 @@ const defaultFields = [
   "payment_status",
   "display_id",
   "cart_id",
+  "draft_order_id",
   "customer_id",
   "email",
   "region_id",