+++ packages/medusa/src/api/routes/admin/shipping-options/create-shipping-option.js
@@ -86,6 +86,7 @@ export default async (req, res) => {
       )
       .optional(),
     is_return: Validator.boolean().default(false),
+    admin_only: Validator.boolean().default(false),
   })
 
   const { value, error } = schema.validate(req.body)