+++ packages/medusa/src/api/routes/admin/shipping-options/index.js
@@ -34,6 +34,7 @@ export const defaultFields = [
   "price_type",
   "amount",
   "is_return",
+  "admin_only",
   "data",
   "created_at",
   "updated_at",