+++ packages/medusa/src/api/routes/admin/shipping-options/update-shipping-option.js
@@ -64,6 +64,7 @@ export default async (req, res) => {
         })
       )
       .optional(),
+    admin_only: Validator.boolean().optional(),
     metadata: Validator.object().optional(),
   })